build.sh -gTtttt on x86 passes ALL tests,examples,benchdnn

+++ build.sh
@@ -802,7 +802,7 @@ if [ "${BUILDOK}" == "y" ]; then
         pwd -P
         ls "${BUILDDIR}/"*log
         for f in "${BUILDDIR}/"*log doxygen.log; do
-            cp -av "${f}" "${LOGDIR}/" || true
+            if [ -f "${f}" ]; then cp -av "${f}" "${LOGDIR}/" || true; fi
         done
     fi
 fi
